(QENG-310) move various host configuration steps together
- remove utils directory, place contents in host_prebuild_steps.rb
- update specs to reflect where code was moved
- update hypervisor.rb to use default, pre-built steps

- addon: remove last reference to systest
