+++ lib/beaker.rb
@@ -1,7 +1,7 @@
 require 'rubygems' unless defined?(Gem)
 module Beaker
 
-  %w( version platform utils test_suite result command options network_manager cli ).each do |lib|
+  %w( version platform test_suite result command options network_manager cli ).each do |lib|
     begin
       require "beaker/#{lib}"
     rescue LoadError
