+++ lib/beaker/hypervisor/vcloud.rb
@@ -7,7 +7,7 @@ module Beaker
     def initialize(vcloud_hosts, options)
       @options = options
       @logger = options[:logger]
-      @vcloud_hosts = vcloud_hosts
+      @hosts = vcloud_hosts
 
       raise 'You must specify a datastore for vCloud instances!' unless @options['datastore']
       raise 'You must specify a resource pool for vCloud instances!' unless @options['resourcepool']
@@ -101,7 +101,7 @@ module Beaker
   
         start = Time.now
         tasks = []
-        @vcloud_hosts.each_with_index do |h, i|
+        @hosts.each_with_index do |h, i|
           # Generate a randomized hostname
           h['vmhostname'] = generate_host_name
   
@@ -135,7 +135,7 @@ module Beaker
   
         try = (Time.now - start) / 5
         duration = run_and_report_duration do 
-          @vcloud_hosts.each_with_index do |h, i|
+          @hosts.each_with_index do |h, i|
             booting_host(h, try, attempts)
           end
         end
@@ -143,7 +143,7 @@ module Beaker
   
         try = (Time.now - start) / 5
         duration = run_and_report_duration do
-          @vcloud_hosts.each_with_index do |h, i|
+          @hosts.each_with_index do |h, i|
             wait_for_dns_resolution(h, try, attempts)
           end
         end
@@ -158,8 +158,8 @@ module Beaker
       @logger.notify "Destroying vCloud boxes"
       connect_to_vsphere
 
-      vm_names = @vcloud_hosts.map {|h| h['vmhostname'] }.compact
-      if @vcloud_hosts.length != vm_names.length
+      vm_names = @hosts.map {|h| h['vmhostname'] }.compact
+      if @hosts.length != vm_names.length
         @logger.warn "Some hosts did not have vmhostname set correctly! This likely means VM provisioning was not successful"
       end
       vms = @vsphere_helper.find_vms vm_names
