+++ lib/beaker/options/presets.rb
@@ -47,6 +47,7 @@ module Beaker
           :timesync => false,
           :repo_proxy => false,
           :add_el_extras => false,
+          :add_master_entry => false,
           :consoleport => 443,
           :pe_dir => '/opt/enterprise/dists',
           :pe_version_file => 'LATEST',
